# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API Gateway serves as middleware between on-premise databases and cloud applications, providing RESTful APIs for ticket and machine management.",
        "title": "API Gateway for On-Premise to Cloud Communication",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "http://www.example.com/license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/machines": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve all machines/terminals from the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Machines"
                ],
                "summary": "Get all machines",
                "responses": {
                    "200": {
                        "description": "List of machines retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.MachineListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/machines/branch/{branch_code}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve all machines for a specific store code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Machines"
                ],
                "summary": "Get machines by store code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store Code",
                        "name": "branch_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Machines retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.MachineListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/machines/metadata": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve all valid values for machine SLM, FLM, NET, and FLM Name fields from the database\nValues are cached for 1 hour for performance. New values automatically appear after cache refresh.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Machines"
                ],
                "summary": "Get machine metadata",
                "responses": {
                    "200": {
                        "description": "Metadata retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.MachineMetadataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/machines/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Search machines using multiple filter criteria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Machines"
                ],
                "summary": "Search machines",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by store code",
                        "name": "store_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by province",
                        "name": "province",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by city/regency",
                        "name": "city_regency",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by district (partial match)",
                        "name": "district",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search completed successfully",
                        "schema": {
                            "$ref": "#/definitions/models.MachineListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/machines/status": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the operational status of a machine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Machines"
                ],
                "summary": "Update machine status",
                "parameters": [
                    {
                        "description": "Machine status update data",
                        "name": "machine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MachineStatusUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Machine status updated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.MachineResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/machines/status/{status}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve all machines with a specific operational status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Machines"
                ],
                "summary": "Get machines by status",
                "parameters": [
                    {
                        "enum": [
                            "Active",
                            "Inactive",
                            "Maintenance",
                            "Offline"
                        ],
                        "type": "string",
                        "description": "Machine Status",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Machines retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.MachineListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/machines/{terminal_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a specific machine by its terminal ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Machines"
                ],
                "summary": "Get machine by terminal ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Terminal ID",
                        "name": "terminal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Machine retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.MachineResponse"
                        }
                    },
                    "404": {
                        "description": "Machine not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ATMI": {
            "type": "object",
            "properties": {
                "city_regency": {
                    "description": "City or regency name",
                    "type": "string",
                    "example": "Jakarta Pusat"
                },
                "date_of_activation": {
                    "description": "Activation date (nullable)",
                    "type": "string",
                    "example": "2023-01-15T00:00:00Z"
                },
                "district": {
                    "description": "District name",
                    "type": "string",
                    "example": "Menteng"
                },
                "gps": {
                    "description": "GPS coordinates",
                    "type": "string",
                    "example": "-6.200000,106.816666"
                },
                "lat": {
                    "description": "Latitude",
                    "type": "number",
                    "example": -6.2
                },
                "lon": {
                    "description": "Longitude",
                    "type": "number",
                    "example": 106.816666
                },
                "province": {
                    "description": "Province name",
                    "type": "string",
                    "example": "DKI Jakarta"
                },
                "status": {
                    "description": "Operational status",
                    "type": "string",
                    "example": "Active"
                },
                "std": {
                    "description": "Standard value",
                    "type": "integer",
                    "example": 1
                },
                "store": {
                    "description": "Store name",
                    "type": "string",
                    "example": "Main Branch"
                },
                "store_code": {
                    "description": "Store code",
                    "type": "string",
                    "example": "STR-001"
                },
                "store_name": {
                    "description": "Full store name",
                    "type": "string",
                    "example": "Jakarta Main Branch"
                },
                "terminal_id": {
                    "description": "Unique terminal identifier",
                    "type": "string",
                    "example": "ATM-001"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Optional detailed error",
                    "type": "string",
                    "example": "detailed error information"
                },
                "message": {
                    "description": "Error message",
                    "type": "string",
                    "example": "Error message describing what went wrong"
                },
                "success": {
                    "description": "Always false for errors",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.FLMInfo": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "string",
                    "example": "BANDUNG"
                },
                "code": {
                    "type": "string",
                    "example": "AVT - BANDUNG"
                },
                "description": {
                    "type": "string",
                    "example": "AVT - BANDUNG"
                },
                "is_documented": {
                    "description": "Whether this value has documentation",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.FLMNameInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "AVT"
                },
                "description": {
                    "type": "string",
                    "example": "AVT"
                },
                "is_documented": {
                    "description": "Whether this value has documentation",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.MachineListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Array of machines",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ATMI"
                    }
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string"
                },
                "success": {
                    "description": "Indicates if operation was successful",
                    "type": "boolean"
                },
                "total": {
                    "description": "Total count of machines",
                    "type": "integer"
                }
            }
        },
        "models.MachineMetadataResponse": {
            "type": "object",
            "properties": {
                "flm_names": {
                    "description": "Available FLM provider names from database",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FLMNameInfo"
                    }
                },
                "flms": {
                    "description": "Available FLM providers from database",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FLMInfo"
                    }
                },
                "last_updated": {
                    "description": "When metadata was last refreshed",
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "message": {
                    "type": "string",
                    "example": "Machine metadata retrieved successfully"
                },
                "nets": {
                    "description": "Available network providers from database",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.NETInfo"
                    }
                },
                "slms": {
                    "description": "Available SLM types from database",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SLMInfo"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.MachineResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Machine data (if applicable)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ATMI"
                        }
                    ]
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string"
                },
                "success": {
                    "description": "Indicates if operation was successful",
                    "type": "boolean"
                }
            }
        },
        "models.MachineStatusUpdate": {
            "type": "object",
            "required": [
                "status",
                "terminal_id"
            ],
            "properties": {
                "gps": {
                    "description": "Optional: GPS coordinates",
                    "type": "string",
                    "example": "-6.200000,106.816666"
                },
                "lat": {
                    "description": "Optional: latitude",
                    "type": "number",
                    "example": -6.2
                },
                "lon": {
                    "description": "Optional: longitude",
                    "type": "number",
                    "example": 106.816666
                },
                "status": {
                    "description": "Required: new status",
                    "type": "string",
                    "example": "Active"
                },
                "terminal_id": {
                    "description": "Required: terminal to update",
                    "type": "string",
                    "example": "ATM-001"
                }
            }
        },
...
        "models.NETInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "NOSAIRIS"
                },
                "description": {
                    "type": "string",
                    "example": "NOSAIRIS"
                },
                "is_documented": {
                    "description": "Whether this value has documentation",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.SLMInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "KGP - WINCOR DW"
                },
                "description": {
                    "type": "string",
                    "example": "KGP - WINCOR DW"
                },
                "is_documented": {
                    "description": "Whether this value has documentation",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for authentication. Required for all endpoints except /health and /ping.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}